+++ anne_auth/app/controllers/anne_auth/application_controller.rb
@@ -1,7 +1,6 @@
 module AnneAuth
   class ApplicationController < ::ApplicationController
     include RouteResolution
-    include Authentication
     include AccountAuthentication
   end
 end