+++ anne_auth/test/generators/install_generator_test.rb
@@ -19,9 +19,8 @@ class AnneAuth::InstallGeneratorTest < Rails::Generators::TestCase
 
     assert_file "config/initializers/anne_auth.rb", /AnneAuth.configure/
     assert_file "config/routes/anne_auth.rb", /mount AnneAuth::Engine/
-    assert_file "app/models/user.rb", /class User < AnneAuth::User/
-    assert_migration "create_anne_auth_users.rb"
-    assert_migration "create_anne_auth_sessions.rb"
+    assert_migration "create_anne_auth_accounts.rb"
+    assert_migration "create_anne_auth_account_sessions.rb"
     assert_migration "create_anne_auth_account_password_reset_tokens.rb"
     assert_no_migration "create_anne_auth_admin_users.rb"
     assert_no_migration "create_anne_auth_admin_sessions.rb"
@@ -31,11 +30,10 @@ class AnneAuth::InstallGeneratorTest < Rails::Generators::TestCase
   test "copies legacy admin migrations when requested" do
     run_generator [ "--legacy-admin" ]
 
+    assert_migration "create_anne_auth_accounts.rb"
+    assert_migration "create_anne_auth_account_sessions.rb"
     assert_migration "create_anne_auth_admin_users.rb"
     assert_migration "create_anne_auth_admin_sessions.rb"
-    assert_no_migration "create_anne_auth_users.rb"
-    assert_no_migration "create_anne_auth_sessions.rb"
-    assert_no_file "app/models/user.rb"
   end
 
   test "does not duplicate migrations when rerun" do