+++ examples/customer_management/app/controllers/application_controller.rb
@@ -1,3 +1,3 @@
 class ApplicationController < ActionController::Base
-  include AnneAuth::Authentication
+  include AnneAuth::AccountAuthentication
 end