+++ examples/customer_management/config/initializers/anne_auth.rb
@@ -1,7 +1,6 @@
 AnneAuth.configure do |config|
   config.mailer_from = "noreply@example.com"
   config.google_oauth_enabled = false
-
-  config.after_login_path = ->(controller, _user) { controller.main_app.admin_root_path }
-  config.after_admin_login_path = ->(controller, user) { config.after_login_path.call(controller, user) }
+  config.account_class_name = "Account"
+  config.after_account_login_path = ->(controller, _account) { controller.main_app.admin_root_path }
 end